Redirect to parameter instead of portal for anonymous accounts
As a workaround we make magic-links redirect to the study instead of the portal, as the portal is not in an adequate state to be able to parse and do the redirects. This will have to be reverted (and given more thought) after the ICAT study is done and there is a need for anonymous deployments for generic web/mobile-hosted studies.